git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/sel-sched-branch@125760 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vect/vect-reduc-dot-u16a.c
@@ -30,7 +30,7 @@ foo1(int len) {
 int main (void)
 {
   unsigned int dot1;
-  int i;
+  unsigned short i;
 
   check_vect ();
 
@@ -46,7 +46,7 @@ int main (void)
   return 0;
 }
 
+/* The initialization loop in main also gets vectorized.  */
 /* { dg-final { scan-tree-dump-times "vect_recog_dot_prod_pattern: detected" 1 "vect" { xfail *-*-* } } } */
-/* { dg-final { scan-tree-dump-times "vectorized 1 loops" 1 "vect" { target { vect_short_mult && vect_widen_sum_hi_to_si } } } } */ 
-
+/* { dg-final { scan-tree-dump-times "vectorized 1 loops" 2 "vect" { target { vect_short_mult && vect_widen_sum_hi_to_si } } } } */ 
 /* { dg-final { cleanup-tree-dump "vect" } } */
